refactor: move common xcm parts to common
